please read description
i took stephen's format and the parts taylor had written and put them in here, as well as added stuff on the theoretical part and the computational approaches. obviously this is far from a finished draft, but it's eough to go on for now that we might start adding more and editing. i'm going to be in austin this weekend but i will turn in what's on here by 11:50 before midnight.